made big search bar for wide screens
