+++ src/pages/ProductsCatalog/styles.module.scss
@@ -57,7 +57,7 @@
 .dropdowns {
   display: flex;
   gap: 16px;
-  margin-bottom: 24px;
+  margin-bottom: 16px;
   width: fit-content;
 }
 
@@ -66,6 +66,18 @@
   font-size: 12px;
   line-height: 15.34px;
   color: $secondary;
+}
+
+.pageSearchBarField {
+  border: 1px solid #b4bdc3;
+  border-radius: 10px;
+  padding-inline: 5px;
+  width: 100%;
+  height: 40px;
+  background-color: white;
+  padding-left: 12px;
+  margin-bottom: 24px;
+  color: hsl(0, 0%, 50%);
 }
 
 .device_list {
